Refactor stock routes
